refactor(#zimic): improve exports
