+++ examples/with-jest-jsdom/tests/interceptors/github.ts
@@ -1,4 +1,4 @@
-import { http } from 'zimic/http/interceptor';
+import { http } from 'zimic/interceptor/http';
 
 import { GITHUB_API_BASE_URL, GitHubRepository } from '../../src/app';
