+++ examples/with-playwright/src/services/github.ts
@@ -1,5 +1,5 @@
 import { cache } from 'react';
-import type { JSONValue } from 'zimic/http';
+import type { JSONValue } from 'zimic';
 
 import environment from '../config/environment';
