+++ examples/with-vitest-browser/src/app.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
-import type { JSONValue } from 'zimic/http';
+import type { JSONValue } from 'zimic';
 
 export const GITHUB_API_BASE_URL = 'https://api.github.com';
