+++ examples/with-vitest-jsdom/src/app.ts
@@ -1,4 +1,4 @@
-import type { JSONValue } from 'zimic/http';
+import type { JSONValue } from 'zimic';
 
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 export const GITHUB_API_BASE_URL = 'https://api.github.com';
