+++ examples/with-vitest-node/src/app.ts
@@ -1,5 +1,5 @@
 import fastify from 'fastify';
-import type { JSONValue } from 'zimic/http';
+import type { JSONValue } from 'zimic';
 import { z } from 'zod';
 
 export const GITHUB_API_BASE_URL = 'https://api.github.com';
