+++ packages/zimic/interceptor/http.d.ts
@@ -1,0 +1,1 @@
+export * from '../dist/interceptor/http';
