+++ packages/zimic/package.json
@@ -52,11 +52,11 @@
       "default": "./dist/http.js",
       "types": "./dist/http.d.ts"
     },
-    "./http/interceptor": {
-      "import": "./dist/http/interceptor.mjs",
-      "require": "./dist/http/interceptor.js",
-      "default": "./dist/http/interceptor.js",
-      "types": "./dist/http/interceptor.d.ts"
+    "./interceptor/http": {
+      "import": "./dist/interceptor/http.mjs",
+      "require": "./dist/interceptor/http.js",
+      "default": "./dist/interceptor/http.js",
+      "types": "./dist/interceptor/http.d.ts"
     },
     "./interceptor/server": {
       "import": "./dist/interceptor/server.mjs",
@@ -87,7 +87,7 @@
     "typegen:fixtures": "tsx ./scripts/dev/typegen/generateFixtureTypes.js",
     "deps:install-playwright": "pnpm playwright install chromium",
     "deps:init-msw": "msw init ./public --no-save",
-    "postinstall": "node -e \"try{require('./dist/scripts/postinstall.js')}catch(error){console.error(error)}\"",
+    "postinstall": "node -e \"try{require('./dist/scripts/postinstall')}catch(error){console.error(error)}\"",
     "prepublish:patch-relative-paths": "sed -E -i 's/\\]\\(\\.\\/([^\\)]+)\\)/](..\\/..\\/\\1)/g;s/\"\\.\\/([^\"]+)\"/\"..\\/..\\/\\1\"/g'",
     "prepublishOnly": "cp ../../README.md ../../LICENSE.md . && pnpm prepublish:patch-relative-paths README.md"
   },
