+++ packages/zimic/src/errors/InvalidJSONError.ts
