+++ packages/zimic/src/index.ts
@@ -1,1 +1,2 @@
-export {};
+export { default as InvalidJSONError } from './errors/InvalidJSONError';
+export type { JSONValue, JSONSerialized } from './types/json';
