+++ packages/zimic/src/typegen/openapi/transform/paths.ts
@@ -43,7 +43,7 @@ function normalizePathNameWithParameters(pathName: string) {
 }
 
 function wrapComponentPathType(type: ts.TypeNode, context: TypeTransformContext) {
-  context.typeImports.root.add('HttpSchema');
+  context.typeImports.http.add('HttpSchema');
 
   const httpSchemaMethodsWrapper = ts.factory.createQualifiedName(
     ts.factory.createIdentifier('HttpSchema'),
@@ -94,7 +94,7 @@ export function normalizePath(
 }
 
 function wrapPathsType(type: ts.TypeLiteralNode, context: TypeTransformContext) {
-  context.typeImports.root.add('HttpSchema');
+  context.typeImports.http.add('HttpSchema');
 
   const httpSchemaPathsWrapper = ts.factory.createQualifiedName(
     ts.factory.createIdentifier('HttpSchema'),
